requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.1
psutil==5.9.8
python-multipart==0.0.6
# 高优先级依赖
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
